sre-crew-chart/requirements.txt
pydantic>=2.4.2
crewai
langchain_openai
litellm[proxy]
kubernetes>=29.0.0
prometheus-api-client
schedule
pyyaml
fastapi
uvicorn